+++ Packs/CommonScripts/Scripts/PrintContext/README.md
@@ -1,0 +1,21 @@
+Pretty-prints the contents of the playbook context.
+
+## Script Data
+---
+
+| **Name** | **Description** |
+| --- | --- |
+| Script Type | python |
+| Tags | Utility |
+
+
+## Inputs
+---
+
+| **Argument Name** | **Description** |
+| --- | --- |
+| outputformat | The output format. Can be, "table", "json" or "markdown". If omitted, the default is "markdown". |
+
+## Outputs
+---
+There are no outputs for this script.
